requirement.txt
aniso8601==3.0.2
click==6.7
Flask==1.0.2
flask-restplus==0.11.0
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.0
PyJWT==1.6.4
pytz==2018.5
six==1.11.0
Werkzeug==0.14.1
